# specs/001-outbound-callcenter-poc/contracts/api.yaml
openapi: 3.0.3
info:
  title: Contact Center POC API
  description: >
    ASP.NET Core Web API for the Outbound Call Center POC.
    Orchestrates outbound phone calls via Azure Communication Services
    and bridges audio to Azure OpenAI Realtime API for AI-driven conversations.
    Supports campaign management, real-time sentiment analysis, multi-number calling,
    and call history with analytics.
  version: 3.0.0
servers:
  - url: https://{host}
    description: API base URL (ngrok tunnel or Azure deployment)
    variables:
      host:
        default: localhost:5001
paths:
...
  /api/Campaign:
    get:
      operationId: listCampaigns
      summary: List all campaigns
      description: >
        Returns all campaigns (pre-defined and custom). Pre-defined campaigns
        have isDefault=true and cannot be deleted.
      tags: [Campaign]
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
    post:
      operationId: createCampaign
      summary: Create a custom campaign
      description: >
        Creates a new custom campaign with the specified title, description,
        and AI behavior instructions. The campaign is persisted to Blob Storage
        and available immediately for call initiation.
      tags: [Campaign]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
            example:
              title: "Product Feedback"
              description: "Collect feedback on recently purchased products"
              aiBehaviorInstructions: "You are a friendly product feedback specialist..."
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Validation error (missing fields or duplicate title)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation error"
                message: "A campaign with this title already exists"
...
  /api/Callback:
    post:
      operationId: callbackEvent
      summary: ACS event callback endpoint
      description: >
        Receives CloudEvent[] callbacks from Azure Communication Services
        for call lifecycle events (CallConnected, CallDisconnected, etc.).
        Not called by the frontend — called by ACS infrastructure.
      tags: [Callback]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                description: CloudEvent from ACS
      responses:
        '200':
          description: Events processed
  /api/Callback/ws:
    get:
      operationId: mediaWebSocket
      summary: ACS media streaming WebSocket
      description: >
        WebSocket endpoint for bidirectional audio streaming between ACS
        and the API. Upgraded from HTTP by ACS when media streaming starts.
        Not called by the frontend — called by ACS infrastructure.
      tags: [Callback]
      parameters:
        - name: targetNumber
          in: query
          required: true
          schema:
            type: string
          description: Target phone number (used to correlate with active call)
      responses:
        '101':
          description: WebSocket upgrade successful
        '400':
          description: Not a WebSocket request
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error category
        message:
          type: string
          description: Human-readable error detail
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          description: Unique campaign identifier (GUID)
        title:
          type: string
          description: Display name
        description:
          type: string
          description: What the campaign is about
        aiBehaviorInstructions:
          type: string
          description: System prompt for the AI agent
        isDefault:
          type: boolean
          description: True for pre-defined campaigns (cannot be deleted)
        createdAt:
          type: string
          format: date-time
    CreateCampaignRequest:
      type: object
      required: [title, description, aiBehaviorInstructions]
      properties:
        title:
          type: string
          description: Campaign display name (must be unique)
        description:
          type: string
          description: What the campaign is about
        aiBehaviorInstructions:
          type: string
          description: System prompt for the AI agent
...